refactor: simplify display of measure using DisplayableMeasure
We're not keeping a reference to the ongoing measure anymore.
